Move clearConntrackEntries to bridge_linux.go
